+++ .gitignore
@@ -25,3 +25,5 @@ java-kanban.iml
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
+
+*.iml
