+++ src/app/service/FileBackedTasksManager.java
@@ -164,21 +164,24 @@ public class FileBackedTasksManager extends InMemoryTaskManager {
     }
 
     @Override
-    public void removeTaskById(int id) {
-        super.removeTaskById(id);
+    public Task removeTaskById(int id) {
+        Task task = super.removeTaskById(id);
         save();
+        return task;
     }
 
     @Override
-    public void removeEpicById(int epicId) {
-        super.removeEpicById(epicId);
+    public Epic removeEpicById(int epicId) {
+        Epic epic = super.removeEpicById(epicId);
         save();
+        return epic;
     }
 
     @Override
-    public void removeSubTaskById(int subtaskIdForRemove) {
-        super.removeSubTaskById(subtaskIdForRemove);
+    public SubTask removeSubTaskById(int subtaskIdForRemove) {
+        SubTask subTask = super.removeSubTaskById(subtaskIdForRemove);
         save();
+        return subTask;
     }
 
     @Override
